requirements/deploy.txt
Django==1.10.7
pytz
Pillow
django-allauth
django-ckeditor
django-imagekit
psycopg2
django-mptt
django-filter
django-nested-admin
